pyproject.toml
[project]
name = "ubuntu-boot-agent"
version = "0.1.0"
description = "AI-powered boot log analysis for Ubuntu Linux"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Daniel Rosehill", email = "public@danielrosehill.com"}
]
dependencies = [
    "httpx>=0.28.1",
    "pyqt6>=6.10.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
ubuntu-boot-agent = "app.ubuntu_boot_agent.__main__:main"

[dependency-groups]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
